Allow setting a non-default audio driver, fixes #466
